внесены изменения
-перенесено создание окна в другой файл
-убрана структура ignore
-коды кнопок заменены дефайнами
