+++ Assert.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <string.h>
+#include "AssertProject/AssertProject/Paint.h"
 
 #ifdef _DEBUG
 #define MyAssert(condition, message) if(!(condition))MyAssertFunc(__LINE__,__FILE__,(message));;
@@ -7,8 +8,6 @@
 #define MyAssert(condition,message)
 #endif // DEBUG
 #define FatalError(condition,message,returnValue) if(!(condition))FatalErrorFunc(__LINE__,__FILE__,(message),(returnValue));;
-
-#pragma warning(disable:4996)
 
 void MyAssertFunc(unsigned int line, char* fileName, char* message);
